+++ calculator/server/sqrt.go
@@ -17,11 +17,11 @@ func (s *Server) Sqrt(ctx context.Context, in *pb.SqrtRequest) (*pb.SqrtResponse
 	if number < 0 {
 		return nil, status.Errorf(
 			codes.InvalidArgument,
-			fmt.Sprintf("Received a negative number: %d", number),
+			fmt.Sprintf("Received a negative number: %f", number),
 		)
 	}
 
 	return &pb.SqrtResponse{
-		Result: math.Sqrt(float64(number)),
+		Result: math.Sqrt(number),
 	}, nil
 }
